# docs/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "API Системы Управления Банковскими Картами"
  description: "REST API для управления банковскими картами с функциями:\n\n- Ауте\
    нтификация и авторизация пользователей\n- Управление картами (создание, просмо\
    тр, блокировка)\n- Переводы между картами\n- Администрирование системы\n\n**Ро\
    ли пользователей:**\n- USER: базовые операции с собственными картами\n- ADMIN:\
    \ полный доступ к системе\n\n**Безопасность:**\n- JWT токены для аутентификаци\
    и\n- Шифрование номеров карт\n- Маскирование чувствительных данных\n"
  contact:
    name: "Abu"
    url: "https://github.com/DesBasito"
    email: "out1of1mind1exception@gmail.com"
  version: "1.0.0"
servers:
- url: "http://localhost:8080/api/v1"
  description: "Локальный сервер разработки"
security:
- Bearer Authentication: []
...
paths:
...
  /transactions/transfer:
    post:
      tags:
      - "Транзакции"
      summary: "Перевод между своими картами"
      description: "Перевод средств между картами текущего пользователя"
      operationId: "transferBetweenMyCards"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "400":
          description: "Некорректные данные для перевода"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "403":
          description: "Карта не принадлежит пользователю"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "200":
          description: "Перевод успешно выполнен"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
      security:
      - Bearer Authentication: []
  /cards/{id}/unblock:
    post:
      tags:
      - "Карты"
      summary: "Отклонить заявку (админ)"
      description: "Отклонение заявки администратором"
      operationId: "unblockCard"
      parameters:
      - name: "id"
        in: "path"
        description: "ID карты"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - Bearer Authentication: []
...
  /transactions/{id}:
    get:
      tags:
      - "Транзакции"
      summary: "Получить транзакцию по ID"
      description: "Получение подробной информации о транзакции"
      operationId: "getTransaction"
      parameters:
      - name: "id"
        in: "path"
        description: "ID транзакции"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Информация о транзакции получена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "404":
          description: "Транзакция не найдена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "403":
          description: "Нет доступа к транзакции"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
      security:
      - Bearer Authentication: []
  /transactions/my:
    get:
      tags:
      - "Транзакции"
      summary: "Получить историю моих транзакций"
      description: "Получение истории транзакций по картам пользователя"
      operationId: "getMyTransactions"
      parameters:
      - name: "cardId"
        in: "query"
        description: "ID карты для фильтрации (необязательно)"
        required: false
        schema:
          type: "integer"
          format: "int64"
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransactionDto"
      security:
      - Bearer Authentication: []
  /transactions/all:
    get:
      tags:
      - "Транзакции"
      summary: "Получить историю транзакций"
      description: "Получение истории транзакций"
      operationId: "getAllTransactions"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransactionDto"
      security:
      - Bearer Authentication: []
  /cards/{id}:
    get:
      tags:
      - "Карты"
      summary: "Получить карту по ID"
      description: "Получение подробной информации о карте"
      operationId: "getCard"
      parameters:
      - name: "id"
        in: "path"
        description: "ID карты"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "Карта не найдена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
        "200":
          description: "Информация о карте получена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
        "403":
          description: "Нет доступа к карте"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - Bearer Authentication: []
  /cards/my:
    get:
      tags:
      - "Карты"
      summary: "Получить все карты пользователя"
      description: "Получение списка карт текущего пользователя с пагинацией"
      operationId: "getMyCards"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "Список карт получен"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
      security:
      - Bearer Authentication: []
  /cards/my/active:
    get:
      tags:
      - "Карты"
      summary: "Получить активные карты пользователя"
      description: "Получение списка активных карт текущего пользователя"
      operationId: "getMyActiveCards"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CardDto"
      security:
      - Bearer Authentication: []
  /cards/all:
    get:
      tags:
      - "Карты"
      summary: "Получить все карты"
      description: "Получение списка карт с пагинацией"
      operationId: "getAllCards"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "Список карт получен"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
...
    TransferRequest:
      required:
      - "amount"
      - "fromCardId"
      - "toCardId"
      type: "object"
      properties:
        fromCardId:
          type: "integer"
          description: "ID карты отправителя"
          format: "int64"
          example: 1
        toCardId:
          type: "integer"
          description: "ID карты получателя"
          format: "int64"
          example: 2
        amount:
          minimum: 0.01
          exclusiveMinimum: false
          type: "number"
          description: "Сумма перевода"
          example: 1500.5
        description:
          maxLength: 500
          minLength: 0
          type: "string"
          description: "Описание перевода"
          example: "Перевод на другую карту"
      description: "Запрос на перевод между картами"
    TransactionDto:
      required:
      - "amount"
      - "createdAt"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID транзакции"
          format: "int64"
        fromCardId:
          type: "integer"
          description: "ID карты отправителя"
          format: "int64"
        fromCardCardNumber:
          type: "string"
          description: "Маскированный номер карты отправителя"
        toCardId:
          type: "integer"
          description: "ID карты получателя"
          format: "int64"
        toCardCardNumber:
          type: "string"
          description: "Маскированный номер карты получателя"
        amount:
          type: "number"
          description: "Сумма транзакции"
        description:
          type: "string"
          description: "Описание транзакции"
        status:
          type: "string"
          description: "Статус транзакции"
          enum:
          - "SUCCESS"
          - "CANCELLED"
          - "REFUNDED"
        createdAt:
          type: "string"
          description: "Дата создания транзакции"
          format: "date-time"
        processedAt:
          type: "string"
          description: "Дата обработки транзакции"
          format: "date-time"
        errorMessage:
          type: "string"
          description: "Сообщение об ошибке (если есть)"
      description: "Информация о транзакции"
    CardDto:
      required:
      - "balance"
      - "cardNumber"
      - "createdAt"
      - "expiryDate"
      - "ownerName"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        cardNumber:
          type: "string"
        ownerName:
          type: "string"
        ownerId:
          type: "integer"
          format: "int64"
        expiryDate:
          type: "string"
          format: "date"
        status:
          type: "string"
        type:
          type: "string"
        balance:
          type: "number"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    Pageable:
      type: "object"
      properties:
        page:
          minimum: 0
          type: "integer"
          format: "int32"
        size:
          minimum: 1
          type: "integer"
          format: "int32"
        sort:
          type: "array"
          items:
            type: "string"
...
    PageableObject:
      type: "object"
      properties:
        offset:
          type: "integer"
          format: "int64"
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: "boolean"
        unpaged:
          type: "boolean"
        pageNumber:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
    SortObject:
      type: "object"
      properties:
        empty:
          type: "boolean"
        unsorted:
          type: "boolean"
        sorted:
          type: "boolean"
...
    PageTransactionDto:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/TransactionDto"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
    Page:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            type: "object"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
...
  securitySchemes:
    Bearer Authentication:
      type: "http"
      description: "Enter JWT token in format: Bearer {token}"
      scheme: "bearer"
      bearerFormat: "JWT"